Move address_map into utils
