+++ tests/test_address_map.C
@@ -1,6 +1,6 @@
 
-#include "address_map.h"
 #include <cyberprobe/network/socket.h>
+#include <cyberprobe/util/address_map.h>
 #include <string>
 #include <assert.h>
